API 추가 - common (공지사항, 문의하기) - member (이메일변경, 비밀번호변경, 회원탈퇴) - report (시선변화 타입 변경)

+++ app/src/main/java/com/sokind/data/remote/report/ReportResponse.kt
@@ -106,7 +106,7 @@ data class ReportEmotion(
     val emotionList: List<Double>,
     val bestEmotion: String,
     @SerializedName("gazeArr")
-    val gazeList: List<Int>,
+    val gazeList: List<Double>,
     val mostGaze: String,
     val gazeImg: String
 )
